requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==42.0.2
Django==5.0.2
django-cors-headers==4.3.1
django-filter==23.5
django-leaflet==0.29.0
django-rest-knox==4.2.0
django-rest-passwordreset==1.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
